// src/app/shared/models/vehicle/vehicle-info.ts
export interface VehicleInfo {
  createdBy: string;
  createdOn: string;
  driverId: string;
  modifiedBy: string;
  modifiedOn: string;
  transporterId: string;
  vehicleId: string;
  vehicleNumber: string;
  vehicleRCLink: string;
  vehicleStatus: string;
  vehicleType: string;
  driverFirstName: string;
  driverLastName: string;
  transporterName: string;
}